Restore changes from BACKLOG-17221
This reverts commit af184f2b9070a5b7cd00b0f8f3cfde7fd1355578.
