+++ src/javascript/SelectorTypes/Picker/PickerDialog/Views/List/SubContentCountLazyLoader.jsx
@@ -27,5 +27,6 @@ export const SubContentCountLazyLoader = ({updateRow, row}) => {
 };
 
 SubContentCountLazyLoader.propTypes = {
-    updateRow: PropTypes.func.isRequired
+    updateRow: PropTypes.func.isRequired,
+    row: PropTypes.object.isRequired
 };
