+++ src/javascript/SelectorTypes/Picker/configs/ContentPickerConfig.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import {Picker} from '../Picker';
 import {File} from '@jahia/moonstone';
 import {useQuery} from '@apollo/react-hooks';
 import {ContentPickerFilledQuery} from './ContentPicker.gql-queries';
@@ -31,24 +30,19 @@ const usePickerInputData = (uuid, editorContext) => {
     return {fieldData, error, loading};
 };
 
-export const ContentPickerSelectorType = {
-    cmp: Picker,
-    key: 'ContentPicker',
-    supportMultiple: false,
+export const ContentPickerConfig = {
     pickerInput: {
         emptyLabel: 'content-editor:label.contentEditor.edit.fields.contentPicker.addContent',
         notFoundLabel: 'content-editor:label.contentEditor.edit.fields.contentPicker.notFoundContent',
         emptyIcon: <File/>,
-        usePickerInputData
+        usePickerInputData,
+        displayTree: true
     },
     PickerDialog: {
         view: 'List',
-        dialogTitle: isPickerTypeFiles => isPickerTypeFiles ?
-            'content-editor:label.contentEditor.edit.fields.contentPicker.modalFileTitle' :
-            'content-editor:label.contentEditor.edit.fields.contentPicker.modalTitle',
-        searchPlaceholder: isPickerTypeFiles => isPickerTypeFiles ?
-            'content-editor:label.contentEditor.edit.fields.contentPicker.searchFilePlaceholder' :
-            'content-editor:label.contentEditor.edit.fields.contentPicker.searchPlaceholder',
-        itemSelectionAdapter: content => content ? content.id : null
+        dialogTitle: 'content-editor:label.contentEditor.edit.fields.contentPicker.modalTitle',
+        searchPlaceholder: 'content-editor:label.contentEditor.edit.fields.contentPicker.searchPlaceholder',
+        itemSelectionAdapter: content => content ? content.id : null,
+        displayTree: true
     }
 };
