+++ src/javascript/SelectorTypes/Picker/configs/Picker.configs.js
@@ -1,0 +1,193 @@
+import {MediaPickerConfig} from '~/SelectorTypes/Picker/configs/MediaPickerConfig';
+import {ContentPickerConfig} from '~/SelectorTypes/Picker/configs/ContentPickerConfig';
+
+const treeConfigs = {
+    content: {
+        getRootPath: site => `/sites/${site}/contents`,
+        openableTypes: ['jnt:contentFolder'],
+        selectableTypes: ['jnt:contentFolder'],
+        type: 'contents',
+        rootLabelKey: 'content-editor:label.contentEditor.edit.fields.contentPicker.contentsRootLabel'
+    },
+    allContents: {
+        getRootPath: site => `/sites/${site}`,
+        openableTypes: ['jnt:page', 'jnt:navMenuText', 'jnt:virtualsite', 'jnt:contentFolder', 'nt:folder', 'jmix:siteContent', 'jmix:browsableInEditorialPicker'],
+        selectableTypes: ['jnt:page', 'jnt:navMenuText', 'jnt:virtualsite', 'jnt:contentFolder', 'nt:folder', 'jmix:siteContent', 'jmix:browsableInEditorialPicker'],
+        type: 'allContents',
+        rootLabelKey: 'content-editor:label.contentEditor.edit.fields.contentPicker.allContentsRootLabel'
+    },
+    files: {
+        getRootPath: site => `/sites/${site}/files`,
+        openableTypes: ['nt:folder'],
+        selectableTypes: ['nt:folder'],
+        type: 'files',
+        rootLabelKey: 'content-editor:label.contentEditor.edit.fields.imagePicker.rootLabel'
+    },
+    pages: {
+        getRootPath: site => `/sites/${site}`,
+        openableTypes: ['jnt:page', 'jnt:virtualsite', 'jnt:navMenuText'],
+        selectableTypes: ['jnt:page', 'jnt:navMenuText'],
+        type: 'pages',
+        rootLabelKey: 'content-editor:label.contentEditor.edit.fields.contentPicker.pagesRootLabel'
+    },
+    categories: {
+        getRootPath: () => '/sites/systemsite/categories',
+        openableTypes: ['jnt:category'],
+        selectableTypes: ['jnt:category'],
+        type: 'categories',
+        rootLabelKey: 'content-editor:label.contentEditor.edit.fields.contentPicker.categoriesRootLabel'
+    },
+    sites: {
+        getRootPath: () => '/sites',
+        openableTypes: ['jnt:virtualsitesFolder'],
+        selectableTypes: ['jnt:virtualsitesFolder'],
+        type: 'sites',
+        rootLabelKey: 'content-editor:label.contentEditor.edit.fields.contentPicker.sitesRootLabel'
+    },
+    users: {
+        getRootPath: () => '/users',
+        openableTypes: ['jnt:usersFolder'],
+        selectableTypes: ['jnt:usersFolder'],
+        type: 'users',
+        rootLabelKey: 'content-editor:label.contentEditor.edit.fields.contentPicker.usersRootLabel'
+    },
+    siteUsers: {
+        getRootPath: site => `/sites/${site}/users`,
+        openableTypes: ['jnt:usersFolder'],
+        selectableTypes: ['jnt:usersFolder'],
+        type: 'users',
+        rootLabelKey: 'content-editor:label.contentEditor.edit.fields.contentPicker.siteUsersRootLabel'
+    },
+    groups: {
+        getRootPath: () => '/groups',
+        openableTypes: ['jnt:groupsFolder'],
+        selectableTypes: ['jnt:groupsFolder'],
+        type: 'groups',
+        rootLabelKey: 'content-editor:label.contentEditor.edit.fields.contentPicker.groupsRootLabel'
+    },
+    siteGroups: {
+        getRootPath: site => `/sites/${site}/groups`,
+        openableTypes: ['jnt:groupsFolder'],
+        selectableTypes: ['jnt:groupsFolder'],
+        type: 'groups',
+        rootLabelKey: 'content-editor:label.contentEditor.edit.fields.contentPicker.siteGroupsRootLabel'
+    }
+};
+const defaultEditorialListType = ['jmix:editorialContent', 'jnt:page', 'jmix:navMenuItem', 'jnt:contentList', 'jnt:contentFolder', 'nt:folder', 'jmix:siteContent'];
+
+export const registerPickerConfig = ceRegistry => {
+    ceRegistry.add('pickerConfiguration', 'editoriallink', {
+        ...ContentPickerConfig,
+        treeConfigs: [treeConfigs.allContents],
+        searchSelectorType: 'jmix:searchable',
+        listTypesTable: defaultEditorialListType,
+        selectableTypesTable: defaultEditorialListType,
+        showOnlyNodesWithTemplates: true
+    });
+
+    ceRegistry.add('pickerConfiguration', 'editorial', {
+        ...ContentPickerConfig,
+        treeConfigs: [treeConfigs.allContents],
+        searchSelectorType: 'jmix:searchable',
+        listTypesTable: defaultEditorialListType,
+        selectableTypesTable: defaultEditorialListType
+    });
+
+    ceRegistry.add('pickerConfiguration', 'image', {
+        ...MediaPickerConfig,
+        treeConfigs: [treeConfigs.files],
+        searchSelectorType: 'jmix:image',
+        listTypesTable: ['jmix:image'],
+        selectableTypesTable: ['jmix:image']
+    });
+
+    ceRegistry.add('pickerConfiguration', 'folder', {
+        ...ContentPickerConfig,
+        pickerInput: {
+            ...ContentPickerConfig.pickerInput,
+            emptyLabel: 'content-editor:label.contentEditor.edit.fields.contentPicker.emptyFolderInput'
+        },
+        PickerDialog: {
+            ...ContentPickerConfig.PickerDialog,
+            dialogTitle: 'content-editor:label.contentEditor.edit.fields.contentPicker.modalFileTitle',
+            searchPlaceholder: 'content-editor:label.contentEditor.edit.fields.contentPicker.searchFilePlaceholder'
+        },
+        treeConfigs: [treeConfigs.files],
+        searchSelectorType: 'jnt:folder',
+        listTypesTable: ['jnt:folder'],
+        selectableTypesTable: ['jnt:folder']
+    });
+
+    ceRegistry.add('pickerConfiguration', 'contentfolder', {
+        ...ContentPickerConfig,
+        PickerDialog: {
+            ...ContentPickerConfig.PickerDialog,
+            dialogTitle: 'content-editor:label.contentEditor.edit.fields.contentPicker.modalFolderTitle'
+        },
+        treeConfigs: [treeConfigs.content],
+        searchSelectorType: 'jnt:contentFolder',
+        listTypesTable: ['jnt:contentFolder'],
+        selectableTypesTable: ['jnt:contentFolder']
+    });
+
+    ceRegistry.add('pickerConfiguration', 'page', {
+        ...ContentPickerConfig,
+        treeConfigs: [treeConfigs.pages],
+        searchSelectorType: 'jnt:page',
+        listTypesTable: ['jnt:page'],
+        selectableTypesTable: ['jnt:page']
+    });
+
+    ceRegistry.add('pickerConfiguration', 'file', {
+        ...ContentPickerConfig,
+        pickerInput: {
+            ...ContentPickerConfig.pickerInput,
+            emptyLabel: 'content-editor:label.contentEditor.edit.fields.contentPicker.addFile'
+        },
+        PickerDialog: {
+            ...ContentPickerConfig.PickerDialog,
+            dialogTitle: 'content-editor:label.contentEditor.edit.fields.contentPicker.modalFileTitle',
+            searchPlaceholder: 'content-editor:label.contentEditor.edit.fields.contentPicker.searchFilePlaceholder'
+        },
+        treeConfigs: [treeConfigs.files],
+        searchSelectorType: 'jnt:file',
+        listTypesTable: ['jnt:file'],
+        selectableTypesTable: ['jnt:file']
+    });
+
+    ceRegistry.add('pickerConfiguration', 'user', {
+        ...ContentPickerConfig,
+        treeConfigs: [treeConfigs.users, treeConfigs.siteUsers],
+        searchSelectorType: 'jnt:user',
+        searchPaths: site => ['/users', `/sites/${site}/users`],
+        listTypesTable: ['jnt:user'],
+        selectableTypesTable: ['jnt:user'],
+        displayTree: false
+    });
+
+    ceRegistry.add('pickerConfiguration', 'usergroup', {
+        ...ContentPickerConfig,
+        treeConfigs: [treeConfigs.groups, treeConfigs.siteGroups],
+        searchSelectorType: 'jnt:group',
+        searchPaths: site => ['/groups', `/sites/${site}/groups`],
+        listTypesTable: ['jnt:group'],
+        selectableTypesTable: ['jnt:group'],
+        displayTree: false
+    });
+
+    ceRegistry.add('pickerConfiguration', 'category', {
+        ...ContentPickerConfig,
+        treeConfigs: [treeConfigs.categories],
+        searchSelectorType: 'jnt:category',
+        listTypesTable: ['jnt:category'],
+        selectableTypesTable: ['jnt:category']
+    });
+
+    ceRegistry.add('pickerConfiguration', 'site', {
+        ...ContentPickerConfig,
+        treeConfigs: [treeConfigs.sites],
+        searchSelectorType: 'jnt:virtualsite',
+        listTypesTable: ['jnt:virtualsite'],
+        selectableTypesTable: ['jnt:virtualsite']
+    });
+};
