+++ src/javascript/SelectorTypes/Picker/index.js
@@ -1,2 +1,3 @@
 export * from './Picker';
+export * from './PickerDialog';
 export * from './registerPicker';
