+++ src/javascript/SelectorTypes/RichText/RichText.utils.spec.js
@@ -1,22 +1,13 @@
 import {buildPickerContext, fillCKEditorPicker} from './RichText.utils';
 
-jest.mock('~/SelectorTypes/Picker/Picker.utils', () => {
-    return {
-        getNodeTreeConfigs: () => ({thisIsANode: 'treeConfig'})
-    };
-});
-
 jest.mock('@jahia/ui-extender', () => {
     return {
         registry: {
             get: jest.fn(() => {
                 return {
-                    cmp: {
-                        picker: {
-                            cmp: {},
-                            key: 'ContentPicker'
-                        }
-                    }
+                    key: 'editorial',
+                    treeConfigs: [{thisIsANode: 'treeConfig'}],
+                    displayTree: true
                 };
             }),
             add: jest.fn()
@@ -26,36 +17,23 @@ jest.mock('@jahia/ui-extender', () => {
 
 describe('RichText utils', () => {
     describe('buildPickerContext', () => {
-        let t;
         let picker;
-        let editorContext;
 
         beforeEach(() => {
-            t = jest.fn();
             picker = {
                 dialog: {
                     getContentElement: jest.fn()
                 }
             };
-            editorContext = {
-                siteInfo: {
-                    displayName: 'digitall'
-                }
-            };
         });
 
         it('should always displayTree', () => {
-            const {pickerConfig} = buildPickerContext(picker, editorContext, t);
+            const {pickerConfig} = buildPickerContext(picker);
             expect(pickerConfig.displayTree).toBe(true);
         });
 
-        it('should return the nodeTreeConfigs', () => {
-            const {nodeTreeConfigs} = buildPickerContext(picker, editorContext, t);
-            expect(nodeTreeConfigs).toEqual({thisIsANode: 'treeConfig'});
-        });
-
         it('should an empty currentValue when input is empty', () => {
-            const {currentValue} = buildPickerContext(picker, editorContext, t);
+            const {currentValue} = buildPickerContext(picker);
             expect(currentValue).toBe('');
         });
 
@@ -65,7 +43,7 @@ describe('RichText utils', () => {
                     getValue: () => '/cms/{mode}/{lang}/richTextEdition/toot/al/regal.html'
                 };
             });
-            const {currentValue} = buildPickerContext(picker, editorContext, t);
+            const {currentValue} = buildPickerContext(picker);
             expect(currentValue).toBe('/richTextEdition/toot/al/regal');
         });
 
@@ -75,7 +53,7 @@ describe('RichText utils', () => {
                     getValue: () => '/files/{workspace}/richTextEdition/toot/al/regal.html'
                 };
             });
-            const {currentValue} = buildPickerContext(picker, editorContext, t);
+            const {currentValue} = buildPickerContext(picker);
             expect(currentValue).toBe('/richTextEdition/toot/al/regal.html');
         });
 
@@ -89,7 +67,7 @@ describe('RichText utils', () => {
                     getValue: () => '/files/{workspace}/richTextEdition/toot/al/regal.html'
                 };
             });
-            const {currentValue} = buildPickerContext(picker, editorContext, t);
+            const {currentValue} = buildPickerContext(picker);
             expect(currentValue).toBe('/richTextEdition/toot/al/regal.html');
         });
     });
